🔥 Refactoring of the code to take out meta util
