+++ plop-templates/component.hbs
@@ -1,6 +1,6 @@
 
 import { useLocation } from 'react-router-dom';
-import { getPlayById } from 'meta/play-meta';
+import { getPlayById } from 'meta/play-meta-util';
 
 import PlayHeader from 'common/playlists/PlayHeader';
