+++ src/common/routing/RouteDefs.js
@@ -1,7 +1,7 @@
 import App from "App";
 import { Footer, Header, Home, PageNotFound } from "common";
 import PlayList from 'common/playlists/PlayList';
-import { getAllPlays } from 'meta/play-meta';
+import { getAllPlays } from 'meta/play-meta-util';
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 
 const RouteDefs = () => {
