+++ src/plays/clock/CurrentTimer.js
@@ -2,7 +2,7 @@
 import PlayHeader from 'common/playlists/PlayHeader';
 import { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
-import { getPlayById } from 'meta/play-meta';
+import { getPlayById } from 'meta/play-meta-util';
 
 const CurrentTimer = () => {
   // The following code is to fetch the current play from the URL
